+++ extensions/broker/src/main/java/sender/message/QueryMessage.java
