+++ extensions/broker/src/main/java/sender/message/RegisterConnectorMessage.java
