+++ extensions/broker/src/main/java/sender/message/RegisterResourceMessage.java
