+++ extensions/broker/src/main/java/transfer/transfer/TransferProcess.java
@@ -1,0 +1,21 @@
+/*
+ *  Copyright (c) 2022 sovity GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       sovity GmbH - initial API and implementation
+ *
+ */
+package transfer.transfer;
+
+import java.net.URL;
+
+public interface TransferProcess {
+
+    void startTransferProcess(URL brokerBaseUrl);
+}
