--- extensions/catalog-transfer-extension/src/main/java/catalogtransfer/transfer/CatalogTransferProcess.java
@@ -1,21 +1,0 @@
-/*
- *  Copyright (c) 2022 sovity GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       sovity GmbH - initial API and implementation
- *
- */
-package catalogtransfer.transfer;
-
-import java.net.URL;
-
-public interface CatalogTransferProcess {
-
-    void startTransferProcess(URL brokerBaseUrl);
-}
