--- extensions/catalog-transfer-extension/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-catalogtransfer.CatalogTransferExtension
