+++ settings.gradle.kts
@@ -1,8 +1,8 @@
-rootProject.name = "broker-extension"
-include("extensions:policies")
-findProject(":extensions:policies")?.name = "policies"
-include("extensions:catalog-transfer-extension")
-findProject(":extensions:catalog-transfer-extension")?.name = "catalog-transfer-extension"
-include("extensions:event-asset-provider")
-findProject(":extensions:event-asset-provider")?.name = "event-asset-provider"
+rootProject.name = "mds-extension"
+include("extensions:broker")
+findProject(":extensions:broker")?.name = "broker"
+include("extensions:clearinghouse")
+findProject(":extensions:clearinghouse")?.name = "clearinghouse"
+include("extensions:event-provider")
+findProject(":extensions:event-provider")?.name = "event-provider"
 include("connector")
